refactor(frontend): shared library map components
Restore the shared library map components for both `BaseMap` and `DataMap` components. App-specific state is passed into each as props.
