+++ frontend/src/app/map/BaseMap.tsx
@@ -1,9 +1,9 @@
 import { MapViewState } from 'deck.gl/typed';
 import { FC, ReactNode } from 'react';
-import { Map } from 'react-map-gl/maplibre';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
 import { mapViewStateState, useSyncMapUrl } from 'app/state/map-view/map-view-state';
+import { BaseMap } from 'lib/data-map/BaseMap';
 import { backgroundState, showLabelsState } from './layers/layers-state';
 import { useBasemapStyle } from './use-basemap-style';
 
@@ -11,11 +11,7 @@ export interface BaseMapProps {
   children?: ReactNode;
 }
 
-/**
- * Displays a react-map-gl basemap component.
- * Accepts children such as a DeckGLOverlay, HUD controls etc
- */
-export const BaseMap: FC<BaseMapProps> = ({ children }) => {
+export const BaseMapContainer: FC<BaseMapProps> = ({ children }) => {
   const background = useRecoilValue(backgroundState);
   const showLabels = useRecoilValue(showLabelsState);
   const [viewState, setViewState] = useRecoilState(mapViewStateState);
@@ -27,20 +23,8 @@ export const BaseMap: FC<BaseMapProps> = ({ children }) => {
   }
 
   return (
-    <Map
-      reuseMaps={true}
-      styleDiffing={true}
-      {...viewState}
-      onMove={handleViewStateChange}
-      mapStyle={mapStyle}
-      dragRotate={false}
-      keyboard={false}
-      touchZoomRotate={true}
-      touchPitch={false}
-      antialias={true}
-      attributionControl={false}
-    >
+    <BaseMap onMove={handleViewStateChange} viewState={viewState} mapStyle={mapStyle}>
       {children}
-    </Map>
+    </BaseMap>
   );
 };
