+++ frontend/src/app/map/MapView.tsx
@@ -1,8 +1,8 @@
 import { Suspense, useCallback } from 'react';
 import { useSetRecoilState } from 'recoil';
 
-import { BaseMap } from './BaseMap';
-import { DataMap } from './DataMap';
+import { BaseMapContainer } from './BaseMap';
+import { DataMapContainer } from './DataMap';
 import { DataMapTooltip } from 'lib/data-map/DataMapTooltip';
 import { MapBoundsFitter, mapFitBoundsState } from 'lib/map/MapBoundsFitter';
 import { MapHud } from 'lib/map/hud/MapHud';
@@ -99,14 +99,14 @@ const MapViewContent = () => {
   const isMobile = useIsMobile();
 
   return (
-    <BaseMap>
-      <DataMap />
+    <BaseMapContainer>
+      <DataMapContainer />
       <MapBoundsFitter />
       <DataMapTooltip>
         <TooltipContent />
       </DataMapTooltip>
       {isMobile ? <MapHudMobileLayout /> : <MapHudDesktopLayout />}
-    </BaseMap>
+    </BaseMapContainer>
   );
 };
